+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "minecraft-text-canvas",
-    "version": "1.1.1",
+    "version": "1.2.0",
     "description": "Generate images of Minecraft text using markup.",
     "license": "MIT",
     "main": "dist/index.js",
