+++ src/lib/set-canvas-dimensions.ts
